Fix #6959: import across sims using madx (#6983)
