+++ sirepo/package_data/static/js/madx.js
@@ -312,7 +312,7 @@ SIREPO.app.directive('appFooter', function() {
         },
         template: `
             <div data-common-footer="nav"></div>
-            <div data-import-dialog="" data-title="Import MAD-X File" data-description="Select a MAD-X file." data-file-formats=".madx,.zip,.seq"></div>
+            <div data-elegant-import-dialog=""></div>
         `,
     };
 });
