+++ sirepo/package_data/static/js/opal.js
@@ -223,9 +223,7 @@ SIREPO.app.directive('appFooter', function() {
         },
         template: `
             <div data-common-footer="nav"></div>
-            <div data-import-dialog="" data-title="Import Opal File" data-description="Select an OPAL .in or .madx file." data-file-formats=".in,.madx,.zip">
-              <div data-import-options=""></div>
-            </div>
+            <div data-elegant-import-dialog=""></div>
         `,
     };
 });
